✨ Add changeConfig cmd (replaces .setsshpassword)
